+++ src/server/products/getProduct.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod';
-import { Product } from '@typings/products/product';
+
 import { ProductSchema } from '@schemas/products/product.schema';
+import type { Product } from '@typings/products/product';
 
 export async function getProduct<K extends 'id' | 'slug'>(field: K, value: Product[K]): Promise<Product | undefined> {
   try {